==> notebooks/01_eda.ipynb <==
import pandas as pd

df_features = pd.read_csv("../elliptic_bitcoin_dataset/elliptic_txs_features.csv", header=None)
df_edges = pd.read_csv("../elliptic_bitcoin_dataset/elliptic_txs_edgelist.csv")
df_classes = pd.read_csv("../elliptic_bitcoin_dataset/elliptic_txs_classes.csv")

print("Features Info: \n")
df_features.info()
print("\n")
print("Classes Info: \n")
df_classes.info()
print("\n")
print("Edge List Info: \n")
df_edges.info()

df_features.head(3)

df_classes.head(3)

df_edges.head(3)

df_classes["class_labeled"] = df_classes["class"].replace({"2": "licit", "1": "illicit"}) 

class_counts = df_classes["class_labeled"].value_counts()
normalized_class_counts = df_classes["class_labeled"].value_counts(normalize=True)

class_analysis_df = pd.DataFrame({
    "Raw Class Names": df_classes["class"].unique(),
    "Labeled Class Names": df_classes["class_labeled"].unique(),
    "Total Counts": class_counts.values,
    "Total Percentage": normalized_class_counts.values 
})
class_analysis_df
